getaddrinfo: Fix leak with AI_ALL [BZ #28852]
Use realloc in convert_hostent_to_gaih_addrtuple and fix up pointers in
the result list so that a single block is maintained for
hostbyname3_r/hostbyname2_r and freed in gaih_inet.  This result is
never merged with any other results, since the hosts database does not
permit merging.

Resolves BZ #28852.

Signed-off-by: Siddhesh Poyarekar <siddhesh@sourceware.org>
Reviewed-by: DJ Delorie <dj@redhat.com>
Upstream-Status: Backport
  https://sourceware.org/git/?p=glibc.git;a=commit;h=300460460706ce3ffe29a7df8966e68323ec5bf1
Signed-off-by: Alex Kiernan <alexk@zuma.ai>